+++ src/Rdd.Domain/Rights/ClosedRightExpressionsHelper.cs
@@ -1,12 +1,13 @@
 ﻿using Rdd.Domain.Models.Querying;
 using System;
 using System.Linq.Expressions;
+using System.Threading.Tasks;
 
 namespace Rdd.Domain.Rights
 {
     public class ClosedRightExpressionsHelper<T> : IRightExpressionsHelper<T>
          where T : class
     {
-        public Expression<Func<T, bool>> GetFilter(Query<T> query) => t => false;
+        public Task<Expression<Func<T, bool>>> GetFilterAsync(Query<T> query) => Task.FromResult<Expression<Func<T, bool>>>(t => false);
     }
 }