+++ src/Rdd.Infra/Storage/ReadOnlyRepository.cs
@@ -23,24 +23,24 @@ namespace Rdd.Infra.Storage
             RightExpressionsHelper = rightExpressionsHelper;
         }
 
-        public virtual Task<int> CountAsync(Query<TEntity> query)
+        public virtual async Task<int> CountAsync(Query<TEntity> query)
         {
             IQueryable<TEntity> entities = Set();
 
             if (query.Options.ChecksRights)
             {
-                entities = ApplyRights(entities, query);
+                entities = await ApplyRightsAsync(entities, query);
             }
 
             entities = ApplyFilters(entities, query);
 
-            return CountEntitiesAsync(entities);
+            return await CountEntitiesAsync(entities);
         }
 
         protected virtual Task<int> CountEntitiesAsync(IQueryable<TEntity> entities)
             => StorageService.CountAsync(entities);
 
-        public virtual Task<IEnumerable<TEntity>> GetAsync(Query<TEntity> query)
+        public virtual async Task<IEnumerable<TEntity>> GetAsync(Query<TEntity> query)
         {
             IQueryable<TEntity> entities = Set();
 
@@ -48,7 +48,7 @@ namespace Rdd.Infra.Storage
 
             if (query.Options.ChecksRights)
             {
-                entities = ApplyRights(entities, query);
+                entities = await ApplyRightsAsync(entities, query);
             }
 
             entities = ApplyFilters(entities, query);
@@ -61,7 +61,7 @@ namespace Rdd.Infra.Storage
 
             entities = ApplyIncludes(entities, query);
 
-            return StorageService.EnumerateEntitiesAsync(entities);
+            return await StorageService.EnumerateEntitiesAsync(entities);
         }
 
         public virtual Task<IEnumerable<TEntity>> PrepareAsync(IEnumerable<TEntity> entities, Query<TEntity> query)
@@ -69,25 +69,26 @@ namespace Rdd.Infra.Storage
             return Task.FromResult(entities);
         }
 
-        public Task<bool> AnyAsync(Query<TEntity> query)
+        public async Task<bool> AnyAsync(Query<TEntity> query)
         {
             IQueryable<TEntity> entities = Set();
 
             if (query.Options.ChecksRights)
             {
-                entities = ApplyRights(entities, query);
+                entities = await ApplyRightsAsync(entities, query);
             }
 
             entities = ApplyFilters(entities, query);
 
-            return StorageService.AnyAsync(entities);
+            return await StorageService.AnyAsync(entities);
         }
 
         protected virtual IQueryable<TEntity> Set() => StorageService.Set<TEntity>();
 
-        protected virtual IQueryable<TEntity> ApplyRights(IQueryable<TEntity> entities, Query<TEntity> query)
+        protected virtual async Task<IQueryable<TEntity>> ApplyRightsAsync(IQueryable<TEntity> entities, Query<TEntity> query)
         {
-            return entities.Where(RightExpressionsHelper.GetFilter(query));
+            var filter = await RightExpressionsHelper.GetFilterAsync(query);
+            return entities.Where(filter);
         }
 
         protected virtual IQueryable<TEntity> ApplyFilters(IQueryable<TEntity> entities, Query<TEntity> query)